Fixed Windows compiler warnings

+++ src/base/RandomWalkComponent.cpp
@@ -94,7 +94,6 @@ bool RandomWalkComponent::GetDirection() {
 }
 
 bool RandomWalkComponent::InBounds(int x, int y) {
-  return (x >= 0 && x < pathGrid.size()) 
-    &&
-    (y >= 0 && y < pathGrid[0].size());
+  return (x >= 0 && (size_t)x < pathGrid.size()) &&
+         (y >= 0 && (size_t)y < pathGrid[0].size());
 }
